requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-Mail==0.10.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
requests==2.32.5
secure-smtplib==0.1.1
urllib3==2.6.3
Werkzeug==3.1.6
